Renamed colours to theme as more accurate
