+++ dot_config/nvim/init-lua.lua
@@ -3,7 +3,7 @@ require 'tap.globals'
 require 'tap.settings'
 require 'tap.plugins'
 require 'tap.diagnostic'
-require 'tap.colours'
+require 'tap.theme'
 require 'tap.auto'
 require 'tap.keymap'
 require 'tap.jest-nvim'
